UI module/requirements.txt
Jinja2==3.1.3
ipython
google-generativeai
gunicorn
easyocr
opencv-python
numpy<2
openai-whisper
ffmpeg-python
pytube
torch
Flask
PyPDF2
pandas
beautifulsoup4
requests
python-dotenv
torchvision
torchaudio
nltk
sentence-transformers
faiss-cpu
streamlit
seaborn
plotly